Add holidays definition (mark holidays)
A bit confusing, since there's a holidays_data.txt already.
Sorry for that.

The holidays_data.txt is the persons' vacation (off-duty) data.

The (new) holidays.yml is the (public) holiday.

Perl stuff:

The join part is a bit tricky. In the beginning, the code was
  class='nameofday dayofweek%d %s'
But that resulted in a blank space before the apostroph for
each (!) day, and many changed lines in the .svg file.

Using join and grep, the blank value (when there's no holiday for
that day) is eliminated by grep, joining only 'nameofday'
and 'dayofweekX'.
Very clean now (only 4 changed lines in the .svg file).